update multi-stream utils
